部分用例增加失败打印信息，18200增加判断规避对象已被删除的问题，17341暂时屏蔽，待问题单4852修复后放开
git-svn-id: http://192.168.20.11/sequoiadb/trunk@42895 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/reliability/java/src/test/java/com/sequoias3/delimiter/DeleteObjectWithKillCoord18200.java
@@ -75,12 +75,14 @@ public class DeleteObjectWithKillCoord18200 extends S3TestBase {
 		Assert.assertEquals(mgr.isAllSuccess(), true, mgr.getErrorMsg());
 
 		//delete again
-		S3Object obj = s3Client.getObject(bucketName, keyName);
-		int versionid = Integer.valueOf(obj.getObjectMetadata().getVersionId());
-		for(int i = versionid ; i >= 0; i--){
-			s3Client.deleteVersion(bucketName, keyName, String.valueOf(i));
+		if(s3Client.doesObjectExist(bucketName, keyName)){
+			S3Object obj = s3Client.getObject(bucketName, keyName);
+			int versionid = Integer.valueOf(obj.getObjectMetadata().getVersionId());
+			for(int i = versionid ; i >= 0; i--){
+				s3Client.deleteVersion(bucketName, keyName, String.valueOf(i));
+			}
+			Assert.assertFalse(s3Client.doesObjectExist(bucketName, keyName));
 		}
-		Assert.assertFalse(s3Client.doesObjectExist(bucketName, keyName));
 
 		runSuccess = true;
 	}
